// types/database.ts
export interface TransactionDB {
    id: string
    patient_name: string
    patient_address: string
    patient_phone: string
    courier_id: string
    prescription_details: string
    total_amount: string
    payment_method: string
    status: string
    notes: string
    created_at: string
    created_by: string
    type: string
    updated_by: string
    updated_at: string
}